수정: 검색바 파일 분류 작업시 코드 수정
