+++ packages/frontend/src/Components/search/FormBox.js
@@ -1,20 +1,10 @@
 import React, { Component, Fragment } from 'react';
 import FormControl from '@material-ui/core/FormControl';
-import Typography from '@material-ui/core/Typography';
-import Paper from '@material-ui/core/Paper';
-import AsyncSelect from 'react-select/async';
-import TextField from '@material-ui/core/TextField';
-import MenuItem from '@material-ui/core/MenuItem';
 import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import { Chip } from '@material-ui/core';
-import CancelIcon from '@material-ui/icons/Cancel';
-
+import Autocomplete from './Autocomplete';
 // import { observer, inject } from "mobx-react";
 // import InputAdornment from '@material-ui/core/InputAdornment';
 // import { MdSearch } from "react-icons/md";
-
-// import Autocomplete from './Autocomplete';
 
 
 const Form = styled.form`
@@ -141,67 +131,6 @@ const Fieldset = styled.fieldset`
 	
 `;
 
-const SelectDiv = styled(AsyncSelect)`
-	&& {
-		// width: 56%;
-		flex-basis: 70%;
-
-		display: inline-flex;
-		position: relative;
-		margin: 16px 0 8px;
-		padding: 0;
-		flex-direction: column;
-		vertical-align: top;
-		transition: all 0.2s ease;
-
-		[placeholder] {
-    		text-overflow:ellipsis;
-  		}  
-
-	}
-`;
-
-const TypographyDiv = styled(Typography)`
-	&& {
-		position: absolute;
-		font-size: 16px;
-	}
-`;
-
-const TypographyNo = styled(Typography)`
-	&& {
-	
-		padding: 14px;
-		font-size: 16px;
-	}
-`;
-
-const ValueDiv = styled.div`
-		display: flex;
-		overflow: hidden;
-		flex-wrap: wrap;
-		flex:1;
-		padding-left: 14px;
-		align-items: center;
-		font-size: 16px;
-
-		& p {
-			font-size: 16px;
-		}
-
-`;
-
-const TextFieldDiv = styled(TextField)`
-	&& {
-		& fieldset + div {
-
-			display: flex;
-			padding: 10px 0 10px;
-			
-		}
-	}
-		
-`;
 
 // 검색아이콘
 // const IconButton = styled.button`
@@ -240,136 +169,12 @@ const TextFieldDiv = styled(TextField)`
 
 // SearchIcon = styled.span``;
 
-const NoOptionsMessage = (props) => {
-	return (
-		<TypographyNo color="textSecondary" className="TypographyNo"
-			{...props.innerProps}>
-			{props.children}
-		</TypographyNo>
-	)
-}
-
-const inputComponent = ({ inputRef, ...props }) => {
-	return < div ref={inputRef} {...props} />
-}
-
-const Control = (props) => {
-	return (
-		<TextFieldDiv fullWidth InputProps={{
-			inputComponent,
-			inputProps: {
-				inputRef: props.innerRef,
-				children: props.children,
-				...props.innerProps,
-			},
-		}}
-			{...props.selectProps.textFieldProps}
-		/>
-	)
-}
-
-const Option = (props) => {
-	return (
-		<MenuItem
-			buttonRef={props.innerRef}
-			selected={props.isFocused}
-			component="div"
-			{...props.innerProps}
-		>
-			{props.children}
-		</MenuItem>
-	)
-}
-
-const Placeholder = (props) => {
-	return (
-		<TypographyDiv
-			color="textSecondary"
-			{...props.innerProps}
-		>
-			{props.children}
-		</TypographyDiv>
-	)
-}
-
-const SingleValue = (props) => {
-	return (
-		<Typography className="" {...props.innerProps}>{props.children}</Typography>
-	)
-}
-
-const ValueContainer = (props) => {
-	return <ValueDiv>{props.children}</ValueDiv>
-}
-
-const MultiValue = (props) => {
-	return (
-		<Chip tabIndex={-1}
-			label={props.children}
-			onDelete={props.removeProps.onClick}
-			deleteIcon={<CancelIcon {...props.removeProps} />}
-		/>
-	)
-}
-
-const Menu = (props) => {
-	return (
-		<Paper square className="" {...props.innerProps}>{props.children}</Paper>
-	)
-}
-
-const components = {
-	Control,
-	Menu,
-	NoOptionsMessage,
-	Option,
-	Placeholder,
-	SingleValue,
-	MultiValue,
-	ValueContainer,
-};
-
-
-
 
 // @inject('searchStore')
 // @observer
 class FormBox extends Component {
 
-	state = {
-		sidoItem: [],
-		single: null,
-		multi: null,
-		error: false
-	}
-
-	// componentDidMount() {
-	// 	this.getAsyncOptions();
-	// }
-
-	handleChange = label => value => {
-		this.setState({
-			[label]: value,
-		});
-
-	};
-
-
 	render() {
-		const getAsyncOptions = () => {
-			const url = `/search/sido`;
-			return fetch(url)
-				.then(response => response.json())
-				.then(json => json.item)
-				.then(data => {
-					console.log(data.map(item => ({ value: item.orgCd, label: item.orgdownNm })));
-					return data.map(item => ({ value: item.orgCd, label: item.orgdownNm }));
-				})
-				.catch(err => {
-					console.log("some error", err.message);
-				});
-		};
-
 		return (
 			<Fragment>
 				{/* <Form autoComplete="off" className={this.props.isVisible ? 'slide-in' : 'slide-out'}> */}
@@ -380,29 +185,7 @@ class FormBox extends Component {
 							{this.props.children}
 						</Fieldset>
 					</FormControlDiv>
-					<SelectDiv
-						inputId="react-select-multiple"
-						textFieldProps={{
-							className: "TextFieldDiv",
-							variant: "outlined",
-							label: '검색어',
-							InputLabelProps: {
-								htmlFor: 'react-select-multiple',
-								shrink: true,
-							}
-						}}
-
-						loadOptions={getAsyncOptions}
-						components={components}
-						value={this.state.multi}
-						onChange={this.handleChange('multi')}
-						// onChange={this.handleChange('single')}
-						placeholder="지역을 입력하세요."
-						isMulti
-						// isClearable
-						cacheOptions
-						defaultOptions
-					/>
+					<Autocomplete />
 					{/* <div className="divider" /> */}
 
 					{/* <TextFieldDiv
@@ -431,138 +214,5 @@ class FormBox extends Component {
 		);
 	}
 }
-
-NoOptionsMessage.propTypes = {
-	/**
-	 * The children to be rendered.
-	 */
-	children: PropTypes.node,
-	/**
-	 * Props to be passed on to the wrapper.
-	 */
-	innerProps: PropTypes.object.isRequired,
-	selectProps: PropTypes.object.isRequired,
-};
-
-inputComponent.propTypes = {
-	inputRef: PropTypes.oneOfType([
-		PropTypes.func,
-		PropTypes.shape({
-			current: PropTypes.any.isRequired,
-		}),
-	]),
-};
-
-Control.propTypes = {
-	/**
-	 * Children to render.
-	 */
-	children: PropTypes.node,
-	/**
-	 * The mouse down event and the innerRef to pass down to the controller element.
-	 */
-	innerProps: PropTypes.shape({
-		onMouseDown: PropTypes.func.isRequired,
-	}).isRequired,
-	innerRef: PropTypes.oneOfType([
-		PropTypes.oneOf([null]),
-		PropTypes.func,
-		PropTypes.shape({
-			current: PropTypes.any.isRequired,
-		}),
-	]).isRequired,
-	selectProps: PropTypes.object.isRequired,
-};
-
-Option.propTypes = {
-	/**
-	 * The children to be rendered.
-	 */
-	children: PropTypes.node,
-	/**
-	 * props passed to the wrapping element for the group.
-	 */
-	innerProps: PropTypes.shape({
-		id: PropTypes.string.isRequired,
-		key: PropTypes.string.isRequired,
-		onClick: PropTypes.func.isRequired,
-		onMouseMove: PropTypes.func.isRequired,
-		onMouseOver: PropTypes.func.isRequired,
-		tabIndex: PropTypes.number.isRequired,
-	}).isRequired,
-	/**
-	 * Inner ref to DOM Node
-	 */
-	innerRef: PropTypes.oneOfType([
-		PropTypes.oneOf([null]),
-		PropTypes.func,
-		PropTypes.shape({
-			current: PropTypes.any.isRequired,
-		}),
-	]).isRequired,
-	/**
-	 * Whether the option is focused.
-	 */
-	isFocused: PropTypes.bool.isRequired,
-	/**
-	 * Whether the option is selected.
-	 */
-	isSelected: PropTypes.bool.isRequired,
-};
-
-Placeholder.propTypes = {
-	/**
-	 * The children to be rendered.
-	 */
-	children: PropTypes.node,
-	/**
-	 * props passed to the wrapping element for the group.
-	 */
-	innerProps: PropTypes.object,
-	selectProps: PropTypes.object.isRequired,
-};
-
-SingleValue.propTypes = {
-	/**
-	 * The children to be rendered.
-	 */
-	children: PropTypes.node,
-	/**
-	 * Props passed to the wrapping element for the group.
-	 */
-	innerProps: PropTypes.any.isRequired,
-	selectProps: PropTypes.object.isRequired,
-};
-
-ValueContainer.propTypes = {
-	/**
-	 * The children to be rendered.
-	 */
-	children: PropTypes.node,
-	selectProps: PropTypes.object.isRequired,
-};
-
-MultiValue.propTypes = {
-	children: PropTypes.node,
-	isFocused: PropTypes.bool.isRequired,
-	removeProps: PropTypes.shape({
-		onClick: PropTypes.func.isRequired,
-		onMouseDown: PropTypes.func.isRequired,
-		onTouchEnd: PropTypes.func.isRequired,
-	}).isRequired,
-	selectProps: PropTypes.object.isRequired,
-};
-
-Menu.propTypes = {
-	/**
-	 * The children to be rendered.
-	 */
-	children: PropTypes.element.isRequired,
-	/**
-	 * Props to be passed to the menu wrapper.
-	 */
-	innerProps: PropTypes.object.isRequired,
-	selectProps: PropTypes.object.isRequired,
-};
 
 export default FormBox;
