+++ yu_report/yu_190805_Search.md
@@ -191,7 +191,12 @@ some() 메서드는 배열 안의 어떤 요소라도 주어진 판별 함수를
   - 첫번째 찾은 검색어 필터링으로 키워드를 찾고 거기서 또 찾고 그러면 필터링을  조건은 && 모두 해당되는 경우만 나오게끔 해야겠다.
   
 * UI변경 자동완성텍스트로 변경 계획
-* 제일먼저 지역(시군구)를 받는다.
+* 제일 먼저 지역(시군구)를 받는다.
 * 착각한부분 api로 받아올 부분을 파라미터로 받다보니 오류가 났다.
-* 시군구를 react-select 라이브러리를 이용하여 fetch로 받아오는 과정에서 async,await을 사용하였는데 계속 안받아와져서 Promise를 이용하니 성공
+* 시군구를 react-select 라이브러리를 이용하여 fetch로 받아오는 과정에서 async,await을 사용하였는데 계속 안받아져서 Promise를 이용하니 성공 (async,await을 이용한 방법을 다시 찾아봐야지.)
+
+* 구현하고 싶은 형태는 검색바 하나에 마우스 포커스되었을때 자동완성텍스트가 나와야하는데 왼쪽에는 시군구 그리고 오른쪽에는 품종이나오고
+	둘 다 선택가능하였으면 좋겠다.
+* 그런데 우선 검색필터 코드가 엄청 길어졌다. 따로 분리하는게 나을거 같다.
+* 그리고 store로 분리 작업도 해야 할듯
